requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiogram-dialog==2.1.0b3
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
arq==0.25.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
Babel==2.12.1
black==23.7.0
blinker==1.6.2
cachetools==5.3.1
cashews==7.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dishka==0.6.0
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.103.1
fastapi-cache2==0.2.1
filelock==3.12.4
Flask==2.3.3
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Dropzone==1.6.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fluent-compiler==0.3
fluent.syntax==0.19.0
fluentogram==1.1.6
fonttools==4.49.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
httpcore==1.0.5
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
magic-filter==1.0.12
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.3
memory-profiler==0.61.0
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.25.2
openpyxl==3.1.2
orjson==3.9.5
packaging==23.1
pandas==2.1.0
passlib==1.7.4
pathspec==0.11.2
pendulum==3.0.0
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.4.0
psutil==5.9.8
psycopg2==2.9.9
py-spy==0.3.14
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
rsa==4.9
selectolax==0.3.16
six==1.16.0
sniffio==1.3.0
speaklater==1.3
SQLAlchemy==2.0.13
starlette==0.27.0
starlette-admin==0.13.2
tablib==3.5.0
text-unidecode==1.3
time-machine==2.13.0
typing_extensions==4.7.1
tzdata==2023.3
uritools==4.0.2
urlextract==1.8.0
urllib3==2.0.6
uvicorn==0.23.2
Werkzeug==2.3.7
WTForms==3.0.1
yarl==1.9.2
